+++ backend/xerox_mfp.c
@@ -49,8 +49,8 @@ dev_request (struct device *dev,
     }
 
     if (cmdlen != len) {
-      DBG (1, "%s: sanei_usb_write_bulk: wanted %d bytes, wrote %d bytes\n",
-	   __FUNCTION__, len, cmdlen);
+      DBG (1, "%s: sanei_usb_write_bulk: wanted %lu bytes, wrote %lu bytes\n",
+	   __FUNCTION__, (u_long)len, (u_long)cmdlen);
       return SANE_STATUS_IO_ERROR;
     }
   }
@@ -118,7 +118,7 @@ static void dbg_dump(struct device *dev)
   for (i = 0; i < dlen; i++, dptr += 3)
     sprintf(dptr, " %02x", dev->res[i]);
 
-  DBG (5, "[%d]%s%s\n", dev->reslen, dbuf,
+  DBG (5, "[%lu]%s%s\n", (u_long)dev->reslen, dbuf,
        (dlen < (int)dev->reslen)? "..." : "");
 }
 
@@ -145,7 +145,8 @@ static int dev_command (struct device *dev, SANE_Byte * cmd, size_t reqlen)
   }
 
   dev->state = 0;
-  DBG (4, ":: dev_command(%s[%#x], %d)\n", str_cmd(cmd[2]), cmd[2], reqlen);
+  DBG (4, ":: dev_command(%s[%#x], %lu)\n", str_cmd(cmd[2]), cmd[2],
+       (u_long)reqlen);
   status = dev_request (dev, cmd, sendlen, res, &dev->reslen);
   if (status != SANE_STATUS_GOOD) {
     DBG (1, "%s: dev_request: %s\n", __FUNCTION__, sane_strstatus (status));
@@ -160,8 +161,8 @@ static int dev_command (struct device *dev, SANE_Byte * cmd, size_t reqlen)
 
   /* normal command reply, some sanity checking */
   if (dev->reslen < reqlen) {
-    DBG (1, "%s: illegal response len %d, need %d\n",
-	 __FUNCTION__, dev->reslen, reqlen);
+    DBG (1, "%s: illegal response len %lu, need %lu\n",
+	 __FUNCTION__, (u_long)dev->reslen, (u_long)reqlen);
     dev->state = SANE_STATUS_IO_ERROR;
     return 0;
   } else {
@@ -177,14 +178,14 @@ static int dev_command (struct device *dev, SANE_Byte * cmd, size_t reqlen)
     }
     pktlen = dev->res[2] + 3;
     if (dev->reslen != pktlen) {
-      DBG (2, "%s: illegal response len %d, should be %d\n",
-	   __FUNCTION__, pktlen, dev->reslen);
+      DBG (2, "%s: illegal response len %lu, should be %lu\n",
+	   __FUNCTION__, (u_long)pktlen, (u_long)dev->reslen);
       dev->state = SANE_STATUS_IO_ERROR;
       return 0;
     }
     if (dev->reslen > reqlen)
-      DBG (2, "%s: too big packet len %d, need %d\n",
-	   __FUNCTION__, dev->reslen, reqlen);
+      DBG (2, "%s: too big packet len %lu, need %lu\n",
+	   __FUNCTION__, (u_long)dev->reslen, (u_long)reqlen);
   }
 
   dev->state = 0;
@@ -1178,15 +1179,15 @@ sane_read (SANE_Handle h, SANE_Byte * buf, SANE_Int maxlen, SANE_Int * lenp)
     while (datalen && dev->blocklen) {
       SANE_Byte *rbuf = dev->data + DATATAIL(dev);
 
-      DBG (9, "<> request len: %d, [%d, %d; %d]\n",
-	   datalen, dev->dataoff, DATATAIL(dev), dev->datalen);
+      DBG (9, "<> request len: %lu, [%d, %d; %d]\n",
+	   (u_long)datalen, dev->dataoff, DATATAIL(dev), dev->datalen);
       if ((status = dev_request (dev, NULL, 0, rbuf, &datalen)) !=
 	  SANE_STATUS_GOOD)
 	return status;
       dev->datalen += datalen;
       dev->blocklen -= datalen;
-      DBG (9, "<> got %d, [%d, %d; %d]\n",
-	   datalen, dev->dataoff, DATATAIL(dev), dev->datalen);
+      DBG (9, "<> got %lu, [%d, %d; %d]\n",
+	   (u_long)datalen, dev->dataoff, DATATAIL(dev), dev->datalen);
       if (dev->blocklen < 0)
 	return ret_cancel(dev, SANE_STATUS_IO_ERROR);
